Refactored shared logic into utils class.
